refactor: split ensemble.py into observability + consensus submodules
Keeps ensemble.py under the 1000-line module limit by extracting the
pure observability linear algebra (sentinel-aware pinvh, observable
subspace bases, the subspace-restricted Mahalanobis distance) into
ensemble_observability.py and the consensus-subset selection (pairwise
agreement matrix, best-neighborhood, consensus_selection) into
ensemble_consensus.py.  No behavior change.
