+++ src/spindoctor/nav_orchestrator/ensemble.py
@@ -38,16 +38,23 @@ import copy
 import math
 from collections.abc import Mapping
 from dataclasses import dataclass, field, fields
-from typing import Any, cast
+from typing import Any
 
 import numpy as np
-from scipy.linalg import pinvh
 
 from spindoctor.annotation import Annotations
 from spindoctor.config import IMAGE_LOGGER
 from spindoctor.feature.constants import (
     AGREEMENT_FACTOR_CAP,
     COMBINED_CONFIDENCE_CAP,
+)
+from spindoctor.nav_orchestrator.ensemble_consensus import (
+    consensus_selection,
+    result_param_vector,
+)
+from spindoctor.nav_orchestrator.ensemble_observability import (
+    mixed_scale_pinvh,
+    observable_basis,
 )
 from spindoctor.nav_orchestrator.feature_summary import NavFeatureSummary
 from spindoctor.nav_orchestrator.image_classifier_result import NavImageClassifierResult
@@ -61,7 +68,7 @@ from spindoctor.nav_technique.nav_technique import technique_tier
 from spindoctor.nav_technique.technique_result import NavTechniqueResult
 from spindoctor.support.exceptions import NavContractError
 from spindoctor.support.status_reason import NavStatusReason
-from spindoctor.support.types import NDArrayBoolType, NDArrayFloatType
+from spindoctor.support.types import NDArrayFloatType
 
 __all__ = [
     'EnsembleConfig',
@@ -93,21 +100,6 @@ DEFAULT_TIER_THRESHOLDS: dict[str, dict[str, float | None]] = {
     'medium': {'min_confidence': 0.2, 'max_sigma_px': 2.0},
     'low': {'min_confidence': 0.2, 'max_sigma_px': None},
 }
-
-# Observability conventions.  A technique reports an unobservable direction
-# in one of two ways: an exactly-zero variance (the Moore-Penrose null-space
-# convention the rank-1 ring-edge path enforces) or a huge
-# axis-aligned sentinel variance (``ROTATION_UNOBSERVABLE_VARIANCE`` = 1e15
-# and the 1e12-scale translation sentinels).  The floor separates genuine
-# uncertainty from sentinels: 1e10 px^2 is a 1e5 px sigma, far beyond any
-# real constraint on a <= 4096 px image and far below the sentinels.  The
-# null threshold matches the rank-1 relative-eigenvalue test used by the
-# ring-edge technique.
-_UNOBSERVABLE_VARIANCE_FLOOR = 1.0e10
-_NULL_RELATIVE_THRESHOLD = 1.0e-8
-# Two projectors' summed eigenvalue must reach 2 (within tolerance) for a
-# direction to lie in the intersection of their subspaces.
-_INTERSECTION_EIGENVALUE_TOL = 1.0e-9
 
 
 @dataclass(frozen=True)
@@ -231,163 +223,6 @@ class EnsembleConfig:
         return cls(**kwargs)
 
 
-def _mixed_scale_pinvh(mat: NDArrayFloatType, *, rcond: float) -> NDArrayFloatType:
-    """Moore-Penrose ``pinvh`` with unobservable-sentinel axes quarantined.
-
-    ``pinvh``'s eigenvalue cutoff is relative to the *largest* eigenvalue,
-    so a huge unobservable-axis sentinel (``ROTATION_UNOBSERVABLE_VARIANCE``
-    or a 1e12-scale translation sentinel) raises the cutoff past every
-    genuine variance: a raw ``pinvh`` on ``diag(0.04, 0.04, 1e15)``
-    truncates the well-constrained 0.04 px^2 axes and the fused offset
-    silently collapses to zero along them.  Sentinel axes are axis-aligned
-    with zero cross-terms by construction, so they are detected by row
-    norm — at or above ``_UNOBSERVABLE_VARIANCE_FLOOR`` (covariance side)
-    or positive and at or below its reciprocal (information side) — and
-    inverted independently as ``1 / diag``; the remaining block gets an
-    ordinary Moore-Penrose ``pinvh``, which the information-form merge's
-    minimum-norm / orthogonal-projection semantics rely on (a merely
-    generalized inverse would project the fused offset obliquely).
-
-    Parameters:
-        mat: Symmetric PSD matrix (covariance or information form).
-        rcond: Relative eigenvalue cutoff for the non-sentinel block.
-
-    Returns:
-        The pseudo-inverse, same shape as ``mat``.
-    """
-    row_norm = np.linalg.norm(mat, axis=1)
-    sentinel = (row_norm >= _UNOBSERVABLE_VARIANCE_FLOOR) | (
-        (row_norm > 0.0) & (row_norm <= 1.0 / _UNOBSERVABLE_VARIANCE_FLOOR)
-    )
-    if not bool(sentinel.any()):
-        return cast(NDArrayFloatType, pinvh(mat, rtol=rcond))
-    out = np.zeros_like(mat)
-    for i in np.flatnonzero(sentinel):
-        out[i, i] = 1.0 / mat[i, i]
-    regular = np.flatnonzero(~sentinel)
-    if regular.size:
-        block = np.ascontiguousarray(mat[np.ix_(regular, regular)])
-        out[np.ix_(regular, regular)] = pinvh(block, rtol=rcond)
-    return out
-
-
-def _observable_basis(cov: NDArrayFloatType) -> NDArrayFloatType:
-    """Orthonormal basis (as columns) of a covariance's observable subspace.
-
-    A direction is unobservable when it is either a sentinel axis (a
-    diagonal variance at or above ``_UNOBSERVABLE_VARIANCE_FLOOR`` — the
-    sentinels are axis-aligned by construction) or a Moore-Penrose null
-    of the remaining block (a relative eigenvalue of the
-    correlation-form matrix below ``_NULL_RELATIVE_THRESHOLD``, so the
-    test is not distorted by mixed px/rad units).
-
-    Parameters:
-        cov: Symmetric PSD covariance, ``(n, n)``.
-
-    Returns:
-        ``(n, k)`` array with orthonormal columns; ``k`` may be zero.
-    """
-    n = cov.shape[0]
-    axes = np.flatnonzero(np.diag(cov) < _UNOBSERVABLE_VARIANCE_FLOOR)
-    if axes.size == 0:
-        return cast(NDArrayFloatType, np.zeros((n, 0), np.float64))
-    sub = cov[np.ix_(axes, axes)]
-    d = np.sqrt(np.clip(np.diag(sub), 0.0, None))
-    d_safe = np.where(d > 0.0, d, 1.0)
-    corr = sub / np.outer(d_safe, d_safe)
-    w, v = np.linalg.eigh(corr)
-    kept = v[:, w > _NULL_RELATIVE_THRESHOLD]
-    if kept.shape[1] == 0:
-        return cast(NDArrayFloatType, np.zeros((n, 0), np.float64))
-    # Map the correlation-space directions back to parameter space
-    # (range(sub) = D @ range(corr)) and re-orthonormalize via SVD.
-    embedded = np.zeros((n, kept.shape[1]), np.float64)
-    embedded[axes, :] = d[:, np.newaxis] * kept
-    u, s, _ = np.linalg.svd(embedded, full_matrices=False)
-    return cast(NDArrayFloatType, u[:, s > _NULL_RELATIVE_THRESHOLD * float(s.max())])
-
-
-def _observable_intersection_basis(
-    cov_a: NDArrayFloatType, cov_b: NDArrayFloatType
-) -> NDArrayFloatType:
-    """Orthonormal basis of the intersection of two observable subspaces.
-
-    Parameters:
-        cov_a: First covariance, ``(n, n)``.
-        cov_b: Second covariance, same shape.
-
-    Returns:
-        ``(n, k)`` array with orthonormal columns; ``k`` may be zero.
-    """
-    basis_a = _observable_basis(cov_a)
-    basis_b = _observable_basis(cov_b)
-    n = cov_a.shape[0]
-    if basis_a.shape[1] == 0 or basis_b.shape[1] == 0:
-        return cast(NDArrayFloatType, np.zeros((n, 0), np.float64))
-    # Eigenvalues of the summed projectors lie in [0, 2]; a direction lies
-    # in both subspaces iff its eigenvalue is 2.
-    proj_sum = basis_a @ basis_a.T + basis_b @ basis_b.T
-    w, v = np.linalg.eigh(proj_sum)
-    return cast(NDArrayFloatType, v[:, w > 2.0 - _INTERSECTION_EIGENVALUE_TOL])
-
-
-def _mahalanobis_distance(
-    mu_a: NDArrayFloatType,
-    cov_a: NDArrayFloatType,
-    mu_b: NDArrayFloatType,
-    cov_b: NDArrayFloatType,
-    *,
-    rcond: float,
-) -> float:
-    """Return the Mahalanobis distance between two estimates.
-
-    Measured only in the intersection of the two estimates' observable
-    subspaces.  A direction one estimate cannot observe
-    carries junk in its parameter vector — the LM/centroid machinery
-    reports *something* there — so differencing it against an estimate
-    that does observe it manufactures disagreement out of nothing (the
-    flat-ring rank-1 result against a full-rank blob), and differencing
-    it against another junk value is meaningless either way.  Estimates
-    with no shared observable direction return 0.0: there is nothing to
-    disagree on, and grouping them lets the information-form combine
-    merge their complementary constraints.
-    """
-    delta = mu_a - mu_b
-    basis = _observable_intersection_basis(cov_a, cov_b)
-    if basis.shape[1] == 0:
-        return 0.0
-    delta_r = basis.T @ delta
-    cov_r = basis.T @ (cov_a + cov_b) @ basis
-    pinv_r = _mixed_scale_pinvh(cov_r, rcond=rcond)
-    d_sq = float(delta_r.T @ pinv_r @ delta_r)
-    if d_sq < 0:
-        # Numerical safety: the generalized inverse may yield a tiny
-        # negative quadratic form due to floating-point; clamp to zero.
-        d_sq = 0.0
-    return float(math.sqrt(d_sq))
-
-
-def _result_param_vector(res: NavTechniqueResult) -> NDArrayFloatType:
-    """Return the parameter vector for a per-technique result.
-
-    Two-DoF results emit ``(dv, du)``; 3-DoF results emit
-    ``(dv, du, rotation_rad)``.  The vector length always matches the
-    covariance shape — a 3x3 covariance with ``rotation_rad=None`` would be
-    inconsistent and raises here.
-    """
-    cov = np.asarray(res.covariance_px2, np.float64)
-    if cov.shape == (3, 3):
-        if res.rotation_rad is None:
-            raise ValueError(
-                f'{res.technique_name}: 3x3 covariance requires rotation_rad to be set'
-            )
-        return cast(
-            NDArrayFloatType,
-            np.array([res.offset_px[0], res.offset_px[1], res.rotation_rad], np.float64),
-        )
-    return cast(NDArrayFloatType, np.array([res.offset_px[0], res.offset_px[1]], np.float64))
-
-
 def _source_bodies(result: NavTechniqueResult) -> frozenset[str]:
     """Return the set of body names a result was computed against.
 
@@ -439,224 +274,6 @@ def _drop_superseded_fallbacks(
                 continue
         kept.append(r)
     return kept
-
-
-@dataclass(frozen=True)
-class _ConsensusSelection:
-    """Output of :func:`_consensus_selection`.
-
-    Parameters:
-        best: The winning consensus subset (never empty).
-        excluded: Viable results outside the winning subset, in input order.
-        runner_up_confidence: Highest summed confidence of any consensus
-            subset formed within ``excluded`` alone; ``0.0`` when nothing
-            was excluded.
-        runner_up_has_quorum: True when that runner-up subset has at least
-            two members -- a genuine alternative consensus rather than a
-            lone dissenter.
-    """
-
-    best: list[NavTechniqueResult]
-    excluded: list[NavTechniqueResult]
-    runner_up_confidence: float
-    runner_up_has_quorum: bool
-
-
-def _agreement_adjacency(
-    results: list[NavTechniqueResult],
-    *,
-    agreement_sigma: float,
-    agreement_pixel_floor: float,
-    rcond: float,
-) -> NDArrayBoolType:
-    """Pairwise agreement matrix: Mahalanobis threshold with a pixel-floor fallback.
-
-    Two results agree iff *either* their pairwise Mahalanobis distance is
-    at most ``agreement_sigma`` *or* their Euclidean translation distance
-    (in pixels, ignoring the rotation component for 3-DoF results) is at
-    most ``agreement_pixel_floor``.
-
-    The Mahalanobis distance differences the rotation component of two
-    3-DoF results linearly; that is correct here because every input
-    rotation is bounded to the small-angle window enforced by the caller,
-    so the pairwise angle difference never approaches the ``+-pi`` wrap.
-
-    The pixel floor compensates for per-technique covariances that
-    report only a CRLB-style precision (FFT subpixel localization or
-    M-estimator information) and so under-estimate the actual position
-    uncertainty by orders of magnitude.  See ``EnsembleConfig`` for the
-    motivation.
-
-    Parameters:
-        results: List of two or more per-technique results.
-        agreement_sigma: Maximum pairwise Mahalanobis distance for agreement.
-        agreement_pixel_floor: Maximum pairwise Euclidean translation
-            distance (in pixels) for agreement; ``0.0`` disables.
-        rcond: rcond passed to ``pinvh``.
-
-    Returns:
-        Symmetric boolean ``(n, n)`` matrix with a True diagonal.
-
-    Raises:
-        ValueError: if 2-DoF and 3-DoF results are mixed.
-    """
-    n = len(results)
-    adj = np.eye(n, dtype=bool)
-    for i in range(n):
-        mu_i = _result_param_vector(results[i])
-        cov_i = np.asarray(results[i].covariance_px2, np.float64)
-        for j in range(i + 1, n):
-            mu_j = _result_param_vector(results[j])
-            cov_j = np.asarray(results[j].covariance_px2, np.float64)
-            if mu_i.shape != mu_j.shape:
-                # 2-DoF and 3-DoF results never coexist within one image.
-                # If the orchestrator routed mismatched shapes through the
-                # ensemble, fail loudly rather than coerce.
-                raise ValueError(
-                    'Mixed-DoF technique results in one ensemble: '
-                    f'{results[i].technique_name} produced {mu_i.shape} parameters; '
-                    f'{results[j].technique_name} produced {mu_j.shape}.'
-                )
-            dist = _mahalanobis_distance(mu_i, cov_i, mu_j, cov_j, rcond=rcond)
-            mahal_match = dist <= agreement_sigma
-            # The pixel floor is likewise measured only in the translation
-            # directions both results observe: a rank-1 result's junk
-            # along-edge component must not push a genuine radial agreement
-            # past the floor.
-            translation_delta = mu_i[:2] - mu_j[:2]
-            basis_t = _observable_intersection_basis(
-                np.ascontiguousarray(cov_i[:2, :2]), np.ascontiguousarray(cov_j[:2, :2])
-            )
-            pixel_dist = float(np.linalg.norm(basis_t.T @ translation_delta))
-            pixel_match = agreement_pixel_floor > 0.0 and pixel_dist <= agreement_pixel_floor
-            if mahal_match or pixel_match:
-                adj[i, j] = True
-                adj[j, i] = True
-    return adj
-
-
-def _best_neighborhood(
-    results: list[NavTechniqueResult],
-    adj: NDArrayBoolType,
-    candidate_indices: list[int],
-) -> list[int]:
-    """The candidate neighborhood with the highest summed confidence.
-
-    Each candidate ``i`` sponsors the subset of ``candidate_indices`` that
-    agree with it pairwise (including itself).  Subsets are scored by
-    summed confidence; ties break toward the larger subset, then toward
-    the earlier candidate, so the selection is deterministic.
-
-    Parameters:
-        results: The full per-technique result list (for confidences).
-        adj: Pairwise agreement matrix over ``results``.
-        candidate_indices: Indices eligible for this selection round.
-
-    Returns:
-        The winning subset as a list of indices into ``results``.
-    """
-    best: list[int] = []
-    best_key = (float('-inf'), 0)
-    for i in candidate_indices:
-        subset = [j for j in candidate_indices if adj[i, j]]
-        key = (sum(results[j].confidence for j in subset), len(subset))
-        if key > best_key:
-            best = subset
-            best_key = key
-    return best
-
-
-def _consensus_selection(
-    results: list[NavTechniqueResult],
-    *,
-    agreement_sigma: float,
-    agreement_pixel_floor: float,
-    rcond: float,
-    max_allowed_rotation_deg: float,
-) -> _ConsensusSelection:
-    """Select the consensus subset the ensemble combines, and its outliers.
-
-    Every result sponsors a candidate subset -- the results that agree
-    with it pairwise (Mahalanobis distance within ``agreement_sigma`` or
-    translation distance within ``agreement_pixel_floor``).  The subset
-    with the highest summed confidence wins; results outside it are
-    excluded from the combine rather than forcing a conflict, and the
-    strongest consensus formable among the excluded results alone is
-    reported as the runner-up so the caller can distinguish a lone
-    mis-converged dissenter from a genuine alternative answer.
-
-    Unlike single-link transitive-closure grouping, a result that agrees
-    with only one member of the winning subset does not drag the whole
-    subset toward it: membership requires pairwise agreement with the
-    sponsoring result.
-
-    Parameters:
-        results: Non-empty list of per-technique results.
-        agreement_sigma: Maximum pairwise Mahalanobis distance for agreement.
-        agreement_pixel_floor: Maximum pairwise Euclidean translation
-            distance (in pixels) for agreement; ``0.0`` disables.
-        rcond: rcond passed to ``pinvh``.
-        max_allowed_rotation_deg: Small-angle bound; each 3-DoF input's
-            ``rotation_rad`` magnitude must stay strictly below this many
-            degrees for the linear rotation differencing to be valid.
-
-    Returns:
-        The selection: winning subset, excluded results, and runner-up
-        summary.
-
-    Raises:
-        NavContractError: if a 3-DoF input's rotation magnitude is at or
-            above the small-angle bound (an upstream programming error).
-        ValueError: if 2-DoF and 3-DoF results are mixed.
-    """
-    n = len(results)
-    max_rotation_rad = math.radians(max_allowed_rotation_deg)
-    for res in results:
-        cov = np.asarray(res.covariance_px2, np.float64)
-        # Small-angle assumption: a 3-DoF rotation outside +-max_rotation_deg
-        # is an upstream error (every technique clamps its rotation fit to
-        # this bound), and would break the linear rotation differencing.
-        # Raised as NavContractError (not a bare assert) so the check
-        # survives python -O and is never swallowed as an ordinary
-        # technique failure.
-        if (
-            cov.shape == (3, 3)
-            and res.rotation_rad is not None
-            and abs(res.rotation_rad) >= max_rotation_rad
-        ):
-            raise NavContractError(
-                f'{res.technique_name}: rotation '
-                f'{math.degrees(res.rotation_rad):.3f} deg violates small-angle '
-                f'bound +-{max_allowed_rotation_deg:.3f} deg'
-            )
-    if n == 1:
-        return _ConsensusSelection(
-            best=list(results),
-            excluded=[],
-            runner_up_confidence=0.0,
-            runner_up_has_quorum=False,
-        )
-    adj = _agreement_adjacency(
-        results,
-        agreement_sigma=agreement_sigma,
-        agreement_pixel_floor=agreement_pixel_floor,
-        rcond=rcond,
-    )
-    best_indices = _best_neighborhood(results, adj, list(range(n)))
-    best_set = set(best_indices)
-    excluded_indices = [i for i in range(n) if i not in best_set]
-    runner_up_confidence = 0.0
-    runner_up_has_quorum = False
-    if excluded_indices:
-        runner_indices = _best_neighborhood(results, adj, excluded_indices)
-        runner_up_confidence = sum(results[j].confidence for j in runner_indices)
-        runner_up_has_quorum = len(runner_indices) >= 2
-    return _ConsensusSelection(
-        best=[results[i] for i in sorted(best_set)],
-        excluded=[results[i] for i in excluded_indices],
-        runner_up_confidence=runner_up_confidence,
-        runner_up_has_quorum=runner_up_has_quorum,
-    )
 
 
 @dataclass(frozen=True)
@@ -739,8 +356,8 @@ def _combine_precision_weighted(
                 f'mixed-DoF group passed to _combine_precision_weighted: '
                 f'expected {n_params}-DoF, got {cov.shape[0]} from {res.technique_name}'
             )
-        info = _mixed_scale_pinvh(cov, rcond=rcond)
-        mu = _result_param_vector(res)
+        info = mixed_scale_pinvh(cov, rcond=rcond)
+        mu = result_param_vector(res)
         if cov.shape[0] == 3:
             theta = float(mu[2])
             # Small-angle assumption: every contributing technique clamps its
@@ -780,7 +397,7 @@ def _combine_precision_weighted(
             info_mu_sum = info_mu_sum + info @ mu
     assert info_sum is not None
     assert info_mu_sum is not None
-    cov_combined = _mixed_scale_pinvh(info_sum, rcond=rcond)
+    cov_combined = mixed_scale_pinvh(info_sum, rcond=rcond)
     mu_combined = cov_combined @ info_mu_sum
     # Rank-deficiency is a property of the translation fix: the combined
     # covariance's (v, u) block has an unobservable direction (exact null
@@ -788,7 +405,7 @@ def _combine_precision_weighted(
     # instead would let an unobservable *rotation* (every input carrying
     # the rotation sentinel) mislabel a fully-constrained 2-D offset as
     # rank-1.
-    trans_basis = _observable_basis(np.ascontiguousarray(cov_combined[:2, :2]))
+    trans_basis = observable_basis(np.ascontiguousarray(cov_combined[:2, :2]))
     is_rank_deficient = trans_basis.shape[1] < 2
     rotation: float | None = None
     if n_params == 3:
@@ -852,7 +469,7 @@ def _combine_confidence(
         # trace (rather than det(info)^(1/p)) stays well-defined for the rank-1
         # ring-edge covariances whose unobservable axis carries zero precision.
         # The 2-DoF path is unchanged: cov[:2, :2] is then the whole matrix.
-        info_xy = _mixed_scale_pinvh(np.ascontiguousarray(cov[:2, :2]), rcond=rcond)
+        info_xy = mixed_scale_pinvh(np.ascontiguousarray(cov[:2, :2]), rcond=rcond)
         weights.append(float(np.trace(info_xy)))
     w_total = sum(weights)
     if w_total <= 0.0:
@@ -1003,7 +620,7 @@ def ensemble(
     interior = [r for r in viable if not r.at_edge]
     if interior:
         viable = interior
-    selection = _consensus_selection(
+    selection = consensus_selection(
         viable,
         agreement_sigma=cfg.agreement_sigma,
         agreement_pixel_floor=cfg.agreement_pixel_floor,
@@ -1135,7 +752,8 @@ def ensemble(
     if rank == 'high' and combined.is_rank_deficient:
         # A fused result with an unobservable translation axis reports one
         # axis as an assumption, not a measurement.  However precise the
-        # observable axis is, that is not a 'high'-tier absolute fix;
+        # observable axis is, that is not a 'high'-tier absolute fix
+        # (issue #221);
         # ``sigma_along_unobservable_px`` carries the same verdict to the
         # metadata.
         IMAGE_LOGGER.info(
